Big re-organization of the `check your answers` html and pdf views.
This PR introduces missing sections like hardship reasons, representative approval, etc.
and also moves around other sections, to match Jason design.

The same, where possible, has been done for `appeal` and `closure` but a following PR will
introduce some missing sections specific to `closure` (as this PR was growing already too much).

* https://www.pivotaltracker.com/story/show/141877591
* https://www.pivotaltracker.com/story/show/141859679
