+++ app/views/steps/closure/check_answers/pdf/show.pdf.erb
@@ -8,7 +8,7 @@
 <body id="case-details-pdf">
 <%= render partial: 'steps/closure/check_answers/pdf/header', locals: { tribunal_case: tribunal_case } %>
 
-<h3><%= pdf_t '.application_details_heading' %></h3>
+<h3><%= t '.taxpayer_details_heading' %></h3>
 
 <hr/>
 <table>
@@ -33,46 +33,7 @@
   </tbody>
 </table>
 
-<hr/>
-<table>
-  <tbody>
-  <tr>
-    <td class="section"><%= pdf_t '.questions.representative_details' %></td>
-    <td class="content">
-      <% if tribunal_case.representative.present? %>
-        <div><%= tribunal_case.representative.name %></div>
-        <div>
-          <%= simple_format(tribunal_case.representative.address, class: 'simple-format') %>
-        </div>
-        <div>
-          <%= pdf_t '.questions.representative_phone' %>
-          <span><%= tribunal_case.representative.phone %></span>
-        </div>
-        <div>
-          <%= pdf_t '.questions.representative_email' %>
-          <span><%= tribunal_case.representative.email %></span>
-        </div>
-      <% else %>
-        <div>
-          <%= pdf_t '.answers.no_representative' %>
-        </div>
-      <% end %>
-    </td>
-  </tr>
-  </tbody>
-</table>
-
-<hr/>
-<table>
-  <tbody>
-  <tr>
-    <td class="section"><%= pdf_t '.questions.legal_representative' %></td>
-    <td class="content">
-      Not yet implemented in data capture.
-    </td>
-  </tr>
-  </tbody>
-</table>
+<%= render partial: 'steps/shared/representative_details', locals: { tribunal_case: tribunal_case } %>
 
 <hr/>
 <table>
