+++ app/views/steps/details/check_answers/pdf/_header.pdf.erb
@@ -1,8 +1,6 @@
 <p>
-  <% if tribunal_case.case_reference.present? %>
-    <span class="case-reference"><%= tribunal_case.case_reference %></span>
-    <br/>
-  <% end %>
-  <strong><%= [tribunal_case.taxpayer.name, 'HMRC'].to_sentence(two_words_connector: t('.appellant_connector')) %></strong>
+  <%= t('.tax_chamber') %>
+  <br/>
+  <strong><%= [t('.notice'), tribunal_case.case_reference].compact.join(' – ') %></strong>
 </p>
 <br/>
