+++ app/views/steps/shared/_representative_details.html.erb
@@ -1,0 +1,49 @@
+<tr>
+  <th><%= t '.questions.representative_details' %></th>
+  <td>
+    <% if @tribunal_case.representative.present? %>
+      <div><%= @tribunal_case.representative.name %></div>
+      <div class="js_breaklines">
+        <%= simple_format(@tribunal_case.representative.address, class: 'simple-format') %>
+      </div>
+      <div>
+        <%= t '.questions.representative_email' %> <span><%= @tribunal_case.representative.email %></span>
+      </div>
+      <div>
+        <%= t '.questions.representative_phone' %> <span><%= @tribunal_case.representative.phone %></span>
+      </div>
+    <% else %>
+      <%= t '.answers.no_representative' %>
+    <% end %>
+  </td>
+  <td class="change-answer">
+    <%= link_to t('.change'), edit_steps_details_has_representative_path %>
+  </td>
+</tr>
+
+<% if tribunal_case.representative.present? %>
+  <tr>
+    <th><%= t '.questions.legal_representative' %></th>
+    <td>
+      <!-- TODO: We only have a Yes/No value at the moment, there is a ticket to change that, this is temporal -->
+      Tax Agent
+    </td>
+    <td class="change-answer">
+      <%= link_to t('.change'), edit_steps_details_representative_is_legal_professional_path %>
+    </td>
+  </tr>
+
+  <% if tribunal_case.representative.proforma_present? %>
+    <tr>
+      <th><%= t '.questions.representative_approval' %></th>
+      <td>
+        <ul class="list-bullet uploaded-docs">
+          <li><%= tribunal_case.representative_approval_file_name %></li>
+        </ul>
+      </td>
+      <td class="change-answer">
+        <%= link_to t('.change'), edit_steps_details_representative_approval_path %>
+      </td>
+    </tr>
+  <% end %>
+<% end %>
