+++ app/views/steps/shared/_representative_details.pdf.erb
@@ -1,0 +1,59 @@
+<hr/>
+<table>
+  <tbody>
+  <tr>
+    <td class="section"><%= pdf_t '.questions.representative_details' %></td>
+    <td class="content">
+      <% if tribunal_case.representative.present? %>
+        <div><%= tribunal_case.representative.name %></div>
+        <div>
+          <%= simple_format(tribunal_case.representative.address, class: 'simple-format') %>
+        </div>
+        <div>
+          <%= pdf_t '.questions.representative_email' %>
+          <span><%= tribunal_case.representative.email %></span>
+        </div>
+        <div>
+          <%= pdf_t '.questions.representative_phone' %>
+          <span><%= tribunal_case.representative.phone %></span>
+        </div>
+      <% else %>
+        <div>
+          <%= pdf_t '.answers.no_representative' %>
+        </div>
+      <% end %>
+    </td>
+  </tr>
+  </tbody>
+</table>
+
+<% if tribunal_case.representative.present? %>
+  <hr/>
+  <table>
+    <tbody>
+    <tr>
+      <td class="section"><%= pdf_t '.questions.legal_representative' %></td>
+      <!-- TODO: We only have a Yes/No value at the moment, there is a ticket to change that, this is temporal -->
+      <td class="content">
+        <span>Tax Agent</span>
+      </td>
+    </tr>
+    </tbody>
+  </table>
+
+  <% if tribunal_case.representative.proforma_present? %>
+    <hr/>
+    <table>
+      <tbody>
+      <tr>
+        <td class="section"><%= pdf_t '.questions.representative_approval' %></td>
+        <td class="content">
+          <ul class="documents">
+            <li><%= tribunal_case.representative_approval_file_name %></li>
+          </ul>
+        </td>
+      </tr>
+      </tbody>
+    </table>
+  <% end %>
+<% end %>
